+++ packages/agtree/src/parser/cosmetic/body/adg-scriptlet.ts
@@ -2,25 +2,33 @@
  * @file AdGuard scriptlet injection body parser
  */
 
+import { sprintf } from 'sprintf-js';
+
 import {
     ADG_SCRIPTLET_MASK,
     CLOSE_PARENTHESIS,
     COMMA,
+    DOUBLE_QUOTE,
     EMPTY,
     ESCAPE_CHARACTER,
     OPEN_PARENTHESIS,
+    SINGLE_QUOTE,
     SPACE,
 } from '../../../utils/constants';
 import { StringUtils } from '../../../utils/string';
 import { AdblockSyntaxError } from '../../../errors/adblock-syntax-error';
 import { ParameterListParser } from '../../misc/parameter-list';
-import { type ScriptletInjectionRuleBody } from '../../common';
 import { defaultParserOptions } from '../../options';
 import { ParserBase } from '../../interface';
 import { type OutputByteBuffer } from '../../../utils/output-byte-buffer';
 import { deserializeScriptletBody, serializeScriptletBody } from './scriptlet-serialization-helper';
 import { type InputByteBuffer } from '../../../utils/input-byte-buffer';
 import { BINARY_SCHEMA_VERSION } from '../../../utils/binary-schema-version';
+import { ValueParser } from '../../misc/value';
+import { type ParameterList, type ScriptletInjectionRuleBody } from '../../common';
+import { isNull } from '../../../utils/type-guards';
+
+type SingleOrDoubleQuote = typeof SINGLE_QUOTE | typeof DOUBLE_QUOTE;
 
 /**
  * `AdgScriptletInjectionBodyParser` is responsible for parsing the body of an AdGuard-style scriptlet rule.
@@ -43,9 +51,12 @@ export class AdgScriptletInjectionBodyParser extends ParserBase {
         NO_SCRIPTLET_MASK: `Invalid ADG scriptlet call, no scriptlet call mask '${ADG_SCRIPTLET_MASK}' found`,
         NO_OPENING_PARENTHESIS: `Invalid ADG scriptlet call, no opening parentheses '${OPEN_PARENTHESIS}' found`,
         NO_CLOSING_PARENTHESIS: `Invalid ADG scriptlet call, no closing parentheses '${CLOSE_PARENTHESIS}' found`,
-        NO_SCRIPTLET_NAME: 'Invalid ADG scriptlet call, no scriptlet name specified',
         WHITESPACE_AFTER_MASK: 'Invalid ADG scriptlet call, whitespace is not allowed after the scriptlet call mask',
         NO_MULTIPLE_SCRIPTLET_CALLS: 'ADG syntaxes does not support multiple scriptlet calls within one single rule',
+        NO_INCONSISTENT_QUOTES: 'Invalid ADG scriptlet call, inconsistent quotes',
+        NO_UNCLOSED_PARAMETER: 'Invalid ADG scriptlet call, unclosed parameter',
+        EXPECTED_QUOTE: "Invalid ADG scriptlet call, expected quote, got '%s'",
+        EXPECTED_COMMA: "Invalid ADG scriptlet call, expected comma, got '%s'",
     };
 
     /**
@@ -181,35 +192,101 @@ export class AdgScriptletInjectionBodyParser extends ParserBase {
             );
         }
 
-        // Parse parameter list
-        const params = ParameterListParser.parse(
-            raw.slice(openingParenthesesIndex + 1, closingParenthesesIndex),
-            options,
-            baseOffset + openingParenthesesIndex + 1,
-            COMMA,
-        );
-
-        // Allow empty scriptlet call: //scriptlet(),
-        // but not allow parameters without scriptlet: //scriptlet(, arg0, arg1)
-        if (params.children.length > 0 && params.children[0] === null) {
-            throw new AdblockSyntaxError(
-                this.ERROR_MESSAGES.NO_SCRIPTLET_NAME,
-                baseOffset + offset,
-                baseOffset + raw.length,
-            );
-        }
+        // Skip space, if any
+        offset = StringUtils.skipWS(raw, offset + 1);
 
         const result: ScriptletInjectionRuleBody = {
             type: 'ScriptletInjectionRuleBody',
-            children: [
-                params,
-            ],
+            children: [],
         };
 
         if (options.isLocIncluded) {
             result.start = baseOffset;
             result.end = baseOffset + raw.length;
         }
+
+        // Special case: empty scriptlet call, like `//scriptlet()`, `//scriptlet( )` etc.
+        if (StringUtils.skipWS(raw, openingParenthesesIndex + 1) === closingParenthesesIndex) {
+            return result;
+        }
+
+        let detectedQuote: SingleOrDoubleQuote | null = null;
+
+        const parameterList: ParameterList = {
+            type: 'ParameterList',
+            children: [],
+        };
+
+        if (options.isLocIncluded) {
+            parameterList.start = baseOffset + openingParenthesesIndex + 1;
+            parameterList.end = baseOffset + closingParenthesesIndex;
+        }
+
+        while (offset < closingParenthesesIndex) {
+            // Skip whitespace
+            offset = StringUtils.skipWS(raw, offset);
+
+            // Expect comma if not first parameter
+            if (parameterList.children.length > 0) {
+                if (raw[offset] !== COMMA) {
+                    throw new AdblockSyntaxError(
+                        sprintf(AdgScriptletInjectionBodyParser.ERROR_MESSAGES.EXPECTED_COMMA, raw[offset]),
+                        baseOffset + offset,
+                        baseOffset + raw.length,
+                    );
+                }
+
+                // Eat the comma
+                offset += 1;
+
+                // Skip whitespace
+                offset = StringUtils.skipWS(raw, offset);
+            }
+
+            // Next character should be a quote
+            if (raw[offset] === SINGLE_QUOTE || raw[offset] === DOUBLE_QUOTE) {
+                if (isNull(detectedQuote)) {
+                    detectedQuote = raw[offset] as SingleOrDoubleQuote;
+                } else if (detectedQuote !== raw[offset]) {
+                    throw new AdblockSyntaxError(
+                        AdgScriptletInjectionBodyParser.ERROR_MESSAGES.NO_INCONSISTENT_QUOTES,
+                        baseOffset + offset,
+                        baseOffset + raw.length,
+                    );
+                }
+
+                // Find next unescaped same quote
+                const closingQuoteIndex = StringUtils.findNextUnescapedCharacter(raw, detectedQuote, offset + 1);
+
+                if (closingQuoteIndex === -1) {
+                    throw new AdblockSyntaxError(
+                        AdgScriptletInjectionBodyParser.ERROR_MESSAGES.NO_UNCLOSED_PARAMETER,
+                        baseOffset + offset,
+                        baseOffset + raw.length,
+                    );
+                }
+
+                // Save the parameter
+                const parameter = ValueParser.parse(
+                    raw.slice(offset, closingQuoteIndex + 1),
+                    options,
+                    baseOffset + offset,
+                );
+
+                parameterList.children.push(parameter);
+
+                // Move after the closing quote
+                offset = StringUtils.skipWS(raw, closingQuoteIndex + 1);
+            } else {
+                throw new AdblockSyntaxError(
+                    sprintf(AdgScriptletInjectionBodyParser.ERROR_MESSAGES.EXPECTED_QUOTE, raw[offset]),
+                    baseOffset + offset,
+                    baseOffset + raw.length,
+                );
+            }
+        }
+
+        result.children.push(parameterList);
 
         return result;
     }